[feat]: 新增 herbiv-cli 及其帮助文档
