+++ .gitignore
@@ -1,3 +1,5 @@
 .idea
 **/__pycache__
 test
+logs
+result
